+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog from v3.0.1 and up. Prior changes don't have a changelog.
 
 * `PersistentPubSubWithOpts` has been removed.
 
+* Rename `Action`'s `Run` method to `Perform`.
+
 # v3.5.1
 
 * Add `EmptyArray` field to `MaybeNil`. (PR #211)